wysyla testowego maila
