+++ .env
@@ -1,6 +1,8 @@
-export SECRET_KEY=gordojestpiekny
+#!/bin/bash
+
 export DEBUG=True
-export APP_SETTINGS=config.DevelopmentConfig
 export DATABASE_URL=sqlite:///database.db
-export FLASK_APP=main.py
+export FLASK_APP="main.py"
 export FLASK_DEBUG=1
+
+source .venv/bin/activate
