+++ pipinstall.txt
@@ -1,18 +1,20 @@
 blinker           
-click             
+click          
 dnspython         
 email_validator   
 Flask             
-Flask-Login       
+Flask-Login
+Flask-Mailman       
+Flask-Migrate
 Flask-SQLAlchemy  
 Flask-WTF         
 greenlet          
 idna             
 itsdangerous      
 Jinja2            
-MarkupSafe        
+MarkupSafe      
+smtplib        
 SQLAlchemy        
 typing_extensions 
 Werkzeug          
-WTForms           
-smtplib
+WTForms
