+++ website/views.py
@@ -1,6 +1,7 @@
 from flask import Blueprint, render_template
 from flask_login import login_required, current_user
 from . import db
+from flask_mailman import EmailMessage
 
 views=Blueprint('views', __name__)
 
@@ -12,3 +13,14 @@ def index():
 @login_required
 def profile():
     return render_template('profile.html', name=current_user.name) #test
+
+@views.route('/testmail')
+def testmail():
+    msg = EmailMessage(
+        'hejhej temat',
+        'hejka tresc',
+        're.click@outlook.com',
+        ['kapi.misiak@gmail.com']
+    )
+    msg.send()
+    return "chyba wyslane"
